ref: Move span processor out from integrations
