+++ sentry_sdk/integrations/opentelemetry/__init__.py
@@ -1,7 +1,5 @@
-from sentry_sdk.integrations.opentelemetry.span_processor import SentrySpanProcessor
 from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
 
 __all__ = [
     "SentryPropagator",
-    "SentrySpanProcessor",
 ]
