+++ sentry_sdk/integrations/opentelemetry/integration.py
@@ -6,10 +6,10 @@ removed at any time without prior notice.
 
 from sentry_sdk.integrations import DidNotEnable, Integration
 from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
-from sentry_sdk.integrations.opentelemetry.span_processor import (
+from sentry_sdk.sampler import SentrySampler
+from sentry_sdk.span_processor import (
     SentrySpanProcessor,
 )
-from sentry_sdk.sampler import SentrySampler
 from sentry_sdk.utils import logger
 
 try:
